Add cut specification to cabinet creation
As part of cabinet dimension specification, the cuts required to create
that cabinet are now generated. The cuts are generated with the type of
woodworking joint used to construct the cabinet.

Cuts are created using final dimensions but internally track dimensions
that consider joinery. Currently, there is no reason for Cuts to track
final dimensions. The Cabinet already tracks this.

For JointType, supporting both DADO and SINGLE_SIDED_DADO because the
back of a cabinet does not join with the top of the cabinet. The
SINGLE_SIDED_DADO implementation seems simpler than tracking the joint
for every side of a cut.