Add panel instance variables to Cabinet, add Dimensions class
Replaced panels/runners instance variables with named instance
variables, for example "side" and "runner". This removed the requirement
for Cut name instance variable. This removal was a response to adding
tests for Cabinet cuts and discovering having direct access to a
particular panel, for example "side", is valuable.

To facilitate testing of Cabinet Cut sizing, added a Dimensions class to
Cut that overrides equals such that a Cut measuring 20 x 30 is equal to a
Cut measuring 30 x 20.

Removed sideMaterial and backMaterial instance variables from Cabinet,
since this information is contained in the Cut class.

Updated style guide to follow more modern java styling convention for
handling overlapping long lines.